pyproject.toml
[project]
name = "pmtk"
version = "0.1.0"
description = "Project Management Toolkit for research."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Conor Bacon", email = "conor.bacon@norsar.no" }
]

dependencies = [
    "typer>=0.12",
    "rich>=13",
    "ruff",
    "pre-commit",
    "pyyaml",
]

[project.scripts]
pmtk = "pmtk.cli:entrypoint"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools.packages.find]
include = ["pmtk*"]

[tool.setuptools.package-data]
pmtk = ["templates/*"]

[tool.ruff]
line-length = 88
